requirements.txt
Django==4.2.7
django-crispy-forms==2.1
crispy-bootstrap5==0.7
Pillow>=10.3.0
WeasyPrint>=60.1
django-environ==0.11.2
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9

